+++ spec/controllers/submissions_controller_spec.rb
@@ -116,6 +116,49 @@ describe SubmissionsController do
       end
     end
 
+    it 'cannot add files outside the repository when an invalid path is given' do
+      file = fixture_file_upload('Shapes.java', 'text/java')
+      bad_path = '../../'
+      post_as @student, :update_files,
+              params: { course_id: course.id, assignment_id: @assignment.id, new_files: [file], path: bad_path }
+
+      expect(response).to have_http_status :bad_request
+      expect(File).to_not exist(File.join(@grouping.group.repo_path, @grouping.assignment.repository_folder,
+                                          bad_path, 'Shapes.java'))
+    end
+
+    it 'cannot add folders outside the repository assignment folder' do
+      dir = '../hello'
+      post_as @student, :update_files,
+              params: { course_id: course.id, assignment_id: @assignment.id, new_folders: [dir], path: '/' }
+
+      expect(response).to have_http_status :unprocessable_entity
+      expect(Dir).to_not exist(File.join(@grouping.group.repo_path, @grouping.assignment.repository_folder,
+                                         dir))
+    end
+
+    it 'cannot delete files outside the repository assignment folder' do
+      post_as @student, :update_files,
+              params: { course_id: course.id, assignment_id: @assignment.id,
+                        delete_files: ['../../../../../LICENSE'], path: '/' }
+
+      expect(response).to have_http_status :unprocessable_entity
+      expect(File).to exist(
+        File.expand_path(File.join(@grouping.group.repo_path, '../../../../../LICENSE'))
+      )
+    end
+
+    it 'cannot delete folders outside the repository assignment folder' do
+      post_as @student, :update_files,
+              params: { course_id: course.id, assignment_id: @assignment.id,
+                        delete_folders: ['../../../../../doc'], path: '/' }
+
+      expect(response).to have_http_status :unprocessable_entity
+      expect(Dir).to exist(
+        File.expand_path(File.join(@grouping.group.repo_path, '../../../../../doc'))
+      )
+    end
+
     context 'submitting a url' do
       describe 'should add url files' do
         before :each do
@@ -1536,6 +1579,23 @@ describe SubmissionsController do
                                                         revision_identifier: submission.revision_identifier }
       end
       it_behaves_like 'notebook types'
+    end
+
+    context 'called with an invalid path' do
+      let(:filename) { 'example.ipynb' }
+      subject do
+        get_as instructor, :notebook_content, params: { course_id: course.id,
+                                                        assignment_id: assignment.id,
+                                                        file_name: filename,
+                                                        grouping_id: grouping.id,
+                                                        revision_identifier: submission.revision_identifier,
+                                                        path: '../..' }
+      end
+
+      it 'flashes an error message' do
+        subject
+        expect(flash[:error].join('\n')).to include(I18n.t('errors.invalid_path'))
+      end
     end
   end